requirements.txt
streamlit
yfinance
plotly
pandas
numpy
SQLAlchemy
scikit-learn
matplotlib
python-dotenv
psycopg2
XlsxWriter
